Rename ctx.rs -> context.rs
